refactor: split entity api routes
