+++ packages/server/src/api/entities.ts
@@ -1,1354 +1,16 @@
-import { randomUUID } from "node:crypto";
 import { Hono } from "hono";
 import type { Kysely } from "kysely";
-import { sql } from "kysely";
-import type { Logger } from "pino";
-import type { Config } from "../config";
-import { type FileViewer, fileVisibilityPredicate } from "../db/repositories/connectors";
-import { createEntityRepository } from "../db/repositories/entities";
-import { type RelationListEntry, createEntityRelationshipsRepository } from "../db/repositories/entity-relationships";
-import { createEntityTimelineRepository } from "../db/repositories/entity-timeline";
 import type { DB } from "../db/schema";
-import { type EntityProfileFacts, SYSTEM_SOURCE_TYPES, mapSourceTypeToEntityType } from "../entities/profile-facts";
-import { getRecreateConflict, recreateEntityGraph } from "../entities/recreate";
-import {
-  beginPendingRebuild,
-  beginRecreateLock,
-  cancelPendingRebuild,
-  endRecreateLock,
-  getPendingRebuild,
-  isRecreateActive,
-  promotePendingRebuild,
-} from "../entities/recreate-state";
-import {
-  AI_EXTRACTION_FACT_TYPES,
-  type ReenrichScope,
-  computeReenrichDryRun,
-  resolveReenrichFileIds,
-  runReenrichJob,
-  tombstoneActiveLlmFacts,
-} from "../entities/reenrich";
-import { denyIfNotAdmin, getFileViewer } from "./auth-helpers";
-import {
-  getCurrentRebuildJob,
-  getCurrentReenrichJob,
-  getCurrentResetJob,
-  getLatestRebuildJob,
-  getLatestReenrichJob,
-  getLatestResetJob,
-  newRebuildJob,
-  newReenrichJob,
-  newResetJob,
-  setCurrentRebuildJob,
-  setCurrentReenrichJob,
-  setCurrentResetJob,
-  setLatestRebuildJob,
-  setLatestReenrichJob,
-  setLatestResetJob,
-} from "./entities/jobs";
-import {
-  FACT_TYPES_BY_CATEGORY,
-  type ResetCategory,
-  computeResetCounts,
-  parseReenrichScope,
-  performReset,
-} from "./entities/reset-service";
+import { createEntityMaintenanceRoutes } from "./entities/maintenance-routes";
+import { createEntityProfileRoutes } from "./entities/profile-routes";
+import type { EntityRoutesDeps } from "./entities/types";
 
 export { _setCurrentReenrichJobForTests, _setCurrentResetJobForTests } from "./entities/jobs";
-
-const RESET_CONFIRM_TOKEN = "RESET_AND_RECREATE";
-const REENRICH_CONFIRM_TOKEN = "REENRICH";
-
-function countByType(rows: RelationListEntry[]): Record<string, number> {
-  const out: Record<string, number> = {};
-  for (const r of rows) {
-    out[r.relationshipType] = (out[r.relationshipType] ?? 0) + 1;
-  }
-  return out;
-}
-
-interface EntityRoutesDeps {
-  logger: Logger;
-  config: Config;
-}
+export type { EntityRoutesDeps } from "./entities/types";
 
 export function entityRoutes(db: Kysely<DB>, deps: EntityRoutesDeps) {
   const routes = new Hono();
-  const repo = createEntityRepository(db);
-  const relRepo = createEntityRelationshipsRepository(db);
-  const timelineRepo = createEntityTimelineRepository(db);
-  const { logger, config } = deps;
-
-  const RELATIONS_LIMIT = 200;
-  const EVIDENCE_LIMIT = 100;
-  const TIMELINE_LIMIT = 100;
-
-  /**
-   * Loads the structured facts shape the drawer Summary block consumes.
-   * Returns null when the entity doesn't exist.
-   */
-  async function loadEntityFactsForId(entityId: string): Promise<EntityProfileFacts | null> {
-    const entity = await repo.getEntity(entityId);
-    if (!entity) return null;
-    const [aggregates, relations] = await Promise.all([
-      repo.getEntityProfileAggregates(entity.id),
-      relRepo.listRelationsForEntity(entity.id, { limit: 50 }),
-    ]);
-    const parsedMetadata = entity.metadata ? (JSON.parse(entity.metadata) as Record<string, unknown>) : null;
-    const topRelationships = [...relations.outgoing, ...relations.incoming].sort(relationCompare).slice(0, 10);
-    return {
-      entityId: entity.id,
-      name: entity.name,
-      sourceType: entity.source_type,
-      entityType: mapSourceTypeToEntityType(entity.source_type),
-      metadata: parsedMetadata,
-      mentionCount: aggregates.mentionCount,
-      sourceCounts: aggregates.sourceCounts,
-      firstSeenAt: aggregates.firstSeenAt,
-      lastSeenAt: aggregates.lastSeenAt,
-      domainsForCompany: aggregates.domainsForCompany,
-      topRelationships,
-      incomingCounts: countByType(relations.incoming),
-      outgoingCounts: countByType(relations.outgoing),
-    };
-  }
-
-  interface ActivityStats {
-    fileCount: number;
-    distinctDays: number;
-    topCoAttendees: string[];
-  }
-
-  /**
-   * Aggregate activity stats for the deterministic Summary: how many files
-   * mention this entity, how many distinct calendar days that spans, and the
-   * top 3 people who co-attend the same files. Two queries, both portable
-   * across SQLite/Postgres. File-derived details are filtered through the
-   * same viewer predicate as timeline/evidence routes.
-   */
-  async function loadActivityStats(entityId: string, viewer: FileViewer): Promise<ActivityStats> {
-    let fileQuery = db
-      .selectFrom("entity_mentions")
-      .innerJoin("indexed_files", "indexed_files.id", "entity_mentions.indexed_file_id")
-      .select([
-        sql<number>`COUNT(DISTINCT entity_mentions.indexed_file_id)`.as("file_count"),
-        sql<number>`COUNT(DISTINCT substr(COALESCE(indexed_files.source_updated_at, indexed_files.source_created_at, entity_mentions.mentioned_at), 1, 10))`.as(
-          "distinct_days",
-        ),
-      ])
-      .where("entity_mentions.entity_id", "=", entityId);
-    if (!viewer.isAdmin) {
-      fileQuery = fileQuery.where(fileVisibilityPredicate(viewer));
-    }
-    const fileRow = await fileQuery.executeTakeFirst();
-
-    let coAttendeeQuery = db
-      .selectFrom("entity_mentions as em1")
-      .innerJoin("indexed_files", "indexed_files.id", "em1.indexed_file_id")
-      .innerJoin("entity_mentions as em2", (join) =>
-        join.onRef("em2.indexed_file_id", "=", "em1.indexed_file_id").on(sql`em2.entity_id <> em1.entity_id`),
-      )
-      .innerJoin("entities as e2", "e2.id", "em2.entity_id")
-      .select(["e2.id as id", "e2.name as name", sql<number>`COUNT(DISTINCT em1.indexed_file_id)`.as("files")])
-      .where("em1.entity_id", "=", entityId)
-      .where("e2.source_type", "=", "person")
-      .groupBy(["e2.id", "e2.name"])
-      .orderBy("files", "desc")
-      .limit(3);
-    if (!viewer.isAdmin) {
-      coAttendeeQuery = coAttendeeQuery.where(fileVisibilityPredicate(viewer));
-    }
-    const coAttendeeRows = await coAttendeeQuery.execute();
-
-    return {
-      fileCount: Number(fileRow?.file_count ?? 0),
-      distinctDays: Number(fileRow?.distinct_days ?? 0),
-      topCoAttendees: coAttendeeRows.map((r) => r.name),
-    };
-  }
-
-  /**
-   * Deterministic prose summary: an identity sentence + an activity sentence.
-   * Built entirely from relationships + aggregates — no LLM, no shimmer. The
-   * idea is to render facts we already know are true rather than synthesizing
-   * narrative from action-item facts.
-   */
-  function buildSummary(facts: EntityProfileFacts, activity: ActivityStats): { identity: string; activity: string } {
-    const identity = buildIdentitySentence(facts);
-    const activitySentence = buildActivitySentence(facts, activity);
-    return { identity, activity: activitySentence };
-  }
-
-  function pluralize(n: number, singular: string, plural?: string): string {
-    return `${n} ${n === 1 ? singular : (plural ?? `${singular}s`)}`;
-  }
-
-  function buildIdentitySentence(facts: EntityProfileFacts): string {
-    const typeWord = formatEntityType(facts.entityType, facts.sourceType);
-    const email = typeof facts.metadata?.email === "string" ? facts.metadata.email : null;
-    const role = typeof facts.metadata?.role === "string" ? facts.metadata.role : null;
-    const outgoing = facts.topRelationships.filter((r) => r.sourceEntityId === facts.entityId);
-    const employers = uniqueNames(outgoing.filter((r) => r.relationshipType === "works_at").map((r) => r.other.name));
-    const clients = uniqueNames(outgoing.filter((r) => r.relationshipType === "engaged_with").map((r) => r.other.name));
-
-    const parts: string[] = [typeWord];
-
-    if (facts.entityType === "person") {
-      if (role) parts.push(role);
-      if (employers.length > 0) parts.push(`works at ${joinList(employers)}`);
-      if (email) parts.push(`(${email})`);
-    } else if (facts.entityType === "company") {
-      const primary = facts.domainsForCompany.find((d) => d.isPrimary)?.domain;
-      const others = facts.domainsForCompany.filter((d) => !d.isPrimary).map((d) => d.domain);
-      if (primary) parts.push(primary);
-      if (others.length > 0) parts.push(`also: ${joinList(others.slice(0, 3))}`);
-    }
-
-    let sentence = parts.join(" · ");
-    if (sentence.length > 0) sentence = `${sentence}.`;
-
-    if (facts.entityType === "person" && clients.length > 0) {
-      sentence = `${sentence} Engaged with ${joinList(clients.slice(0, 4))}.`;
-    }
-
-    return sentence;
-  }
-
-  function buildActivitySentence(facts: EntityProfileFacts, activity: ActivityStats): string {
-    const pieces: string[] = [];
-    if (activity.fileCount > 0) {
-      const fileTerm = pluralize(activity.fileCount, "file");
-      const mentionsTerm = pluralize(facts.mentionCount, "mention");
-      if (activity.distinctDays > 0) {
-        pieces.push(`Active in ${fileTerm} (${mentionsTerm}) across ${pluralize(activity.distinctDays, "day")}`);
-      } else {
-        pieces.push(`Active in ${fileTerm} (${mentionsTerm})`);
-      }
-    }
-    if (facts.entityType === "person" && activity.topCoAttendees.length > 0) {
-      pieces.push(`most-frequent collaborators: ${joinList(activity.topCoAttendees)}`);
-    }
-    if (pieces.length === 0) return "";
-    return `${pieces.join(". ")}.`;
-  }
-
-  function formatEntityType(entityType: string, sourceType: string): string {
-    switch (entityType) {
-      case "person":
-        return "Person";
-      case "company":
-        return "Company";
-      case "product":
-        return "Product";
-      case "project":
-        return "Project";
-      case "team":
-        return "Team";
-      case "system":
-        return capitalize(sourceType);
-      default:
-        return capitalize(sourceType);
-    }
-  }
-
-  function capitalize(value: string): string {
-    if (!value) return value;
-    return value.charAt(0).toUpperCase() + value.slice(1);
-  }
-
-  function uniqueNames(names: string[]): string[] {
-    const seen = new Set<string>();
-    const out: string[] = [];
-    for (const name of names) {
-      const trimmed = name.trim();
-      if (!trimmed) continue;
-      const key = trimmed.toLowerCase();
-      if (seen.has(key)) continue;
-      seen.add(key);
-      out.push(trimmed);
-    }
-    return out;
-  }
-
-  function joinList(items: string[]): string {
-    if (items.length === 0) return "";
-    if (items.length === 1) return items[0];
-    if (items.length === 2) return `${items[0]} and ${items[1]}`;
-    return `${items.slice(0, -1).join(", ")}, and ${items[items.length - 1]}`;
-  }
-  const CONFIDENCE_ORDER: Record<string, number> = { AMBIGUOUS: 0, EXTRACTED: 1, INFERRED: 2 };
-  const relationCompare = (a: RelationListEntry, b: RelationListEntry): number => {
-    const ac = CONFIDENCE_ORDER[a.confidence] ?? 3;
-    const bc = CONFIDENCE_ORDER[b.confidence] ?? 3;
-    if (ac !== bc) return ac - bc;
-    if (a.evidenceCount !== b.evidenceCount) return b.evidenceCount - a.evidenceCount;
-    if (a.confidenceScore !== b.confidenceScore) return b.confidenceScore - a.confidenceScore;
-    const byName = a.other.name.localeCompare(b.other.name);
-    if (byName !== 0) return byName;
-    return a.id.localeCompare(b.id);
-  };
-
-  /**
-   * GET /api/entities
-   *   ?type=person,clickup_space     — filter by source_type (comma-separated)
-   *   &source=clickup,linear          — filter by source (from entity_source_refs)
-   *   &search=beetu                   — name/alias search
-   *   &sort=hotness|mentions|name     — sort field (default: hotness)
-   *   &limit=50&offset=0
-   */
-  /**
-   * POST /api/entities
-   * Create a new entity manually.
-   */
-  routes.post("/", async (c) => {
-    const body = (await c.req.json()) as {
-      name: string;
-      sourceType: string;
-      subtype?: string;
-      aliases?: string[];
-    };
-
-    if (!body.name?.trim() || !body.sourceType?.trim()) {
-      return c.json({ error: { code: "BAD_REQUEST", message: "name and sourceType are required" } }, 400);
-    }
-
-    const entity = await repo.upsertEntity({
-      name: body.name.trim(),
-      sourceType: body.sourceType.trim(),
-      subtype: body.subtype,
-      aliases: body.aliases,
-      status: "confirmed",
-    });
-
-    return c.json({
-      entity: {
-        id: entity.id,
-        name: entity.name,
-        sourceType: entity.source_type,
-        subtype: entity.subtype,
-        aliases: entity.aliases ? JSON.parse(entity.aliases) : [],
-        status: entity.status,
-      },
-    });
-  });
-
-  routes.get("/", async (c) => {
-    const typeFilter = c.req.query("type")?.split(",").filter(Boolean);
-    const sourceFilter = c.req.query("source")?.split(",").filter(Boolean);
-    const search = c.req.query("search");
-    const sort = c.req.query("sort") ?? "hotness";
-    const limit = Math.min(Number(c.req.query("limit")) || 50, 200);
-    const offset = Number(c.req.query("offset")) || 0;
-    const includeSystem = c.req.query("includeSystem") === "true";
-    const includeArchived = c.req.query("includeArchived") === "true";
-    const systemTypes = [...SYSTEM_SOURCE_TYPES];
-
-    let query = db
-      .selectFrom("entities")
-      .selectAll("entities")
-      .select(
-        sql<number>`(SELECT count(*) FROM entity_mentions WHERE entity_mentions.entity_id = entities.id)`.as(
-          "mention_count",
-        ),
-      )
-      .select(
-        sql<string>`(SELECT COALESCE(indexed_files.source_updated_at, indexed_files.source_created_at, entity_mentions.mentioned_at) FROM entity_mentions INNER JOIN indexed_files ON indexed_files.id = entity_mentions.indexed_file_id WHERE entity_mentions.entity_id = entities.id ORDER BY COALESCE(indexed_files.source_updated_at, indexed_files.source_created_at, entity_mentions.mentioned_at) DESC LIMIT 1)`.as(
-          "last_mention_at",
-        ),
-      );
-
-    if (typeFilter && typeFilter.length > 0) {
-      query = query.where("entities.source_type", "in", typeFilter);
-    }
-
-    if (sourceFilter && sourceFilter.length > 0) {
-      query = query.where(
-        "entities.id",
-        "in",
-        db.selectFrom("entity_source_refs").select("entity_id").where("source", "in", sourceFilter),
-      );
-    }
-
-    if (search) {
-      const pattern = `%${search}%`;
-      query = query.where((eb) =>
-        eb.or([eb("entities.name", "like", pattern), eb("entities.aliases", "like", pattern)]),
-      );
-    }
-
-    if (!includeArchived) {
-      query = query.where("entities.status", "!=", "archived");
-    }
-
-    // Hide container/system entities (clickup_workspace, clickup_space) unless
-    // the caller opts in. Explicit type filters always win — deep links and the
-    // "Show system entities" toggle keep working.
-    if (!typeFilter && !includeSystem && systemTypes.length > 0) {
-      query = query.where("entities.source_type", "not in", systemTypes);
-    }
-
-    if (sort === "mentions") {
-      query = query.orderBy("mention_count", "desc");
-    } else if (sort === "name") {
-      query = query.orderBy("entities.name", "asc");
-    } else {
-      query = query.orderBy("entities.hotness", "desc");
-    }
-
-    query = query.limit(limit).offset(offset);
-
-    const entities = await query.execute();
-
-    // Total count for pagination
-    let countQuery = db.selectFrom("entities").select(db.fn.count("id").as("total"));
-    if (typeFilter && typeFilter.length > 0) {
-      countQuery = countQuery.where("source_type", "in", typeFilter);
-    }
-    if (search) {
-      const pattern = `%${search}%`;
-      countQuery = countQuery.where((eb) => eb.or([eb("name", "like", pattern), eb("aliases", "like", pattern)]));
-    }
-    if (!includeArchived) {
-      countQuery = countQuery.where("status", "!=", "archived");
-    }
-    if (!typeFilter && !includeSystem && systemTypes.length > 0) {
-      countQuery = countQuery.where("source_type", "not in", systemTypes);
-    }
-    const countResult = await countQuery.executeTakeFirst();
-
-    return c.json({
-      entities: entities.map((e) => ({
-        id: e.id,
-        name: e.name,
-        sourceType: e.source_type,
-        subtype: e.subtype,
-        aliases: e.aliases ? JSON.parse(e.aliases) : [],
-        metadata: e.metadata ? JSON.parse(e.metadata) : null,
-        status: e.status,
-        hotness: e.hotness,
-        mentionCount: Number(e.mention_count ?? 0),
-        lastMentionAt: e.last_mention_at ?? null,
-        createdAt: e.created_at,
-        updatedAt: e.updated_at,
-      })),
-      total: Number(countResult?.total ?? 0),
-    });
-  });
-
-  const ORG_SOURCE_TYPES = ["person", "company", "product", "team", "project"];
-
-  /**
-   * GET /api/entities/resets/jobs
-   * Current + most-recently-finished reset+rebuild job.
-   */
-  routes.get("/resets/jobs", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const conflict = await getRecreateConflict(db);
-    const currentResetJob = getCurrentResetJob();
-    return c.json({
-      active: currentResetJob !== null || isRecreateActive(),
-      currentJob: currentResetJob,
-      latestJob: getLatestResetJob(),
-      blockedBy: currentResetJob ? null : conflict,
-    });
-  });
-
-  /**
-   * GET /api/entities/resets/jobs/:id
-   */
-  routes.get("/resets/jobs/:id", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const id = c.req.param("id");
-    const currentResetJob = getCurrentResetJob();
-    const latestResetJob = getLatestResetJob();
-    if (currentResetJob?.id === id) return c.json(currentResetJob);
-    if (latestResetJob?.id === id) return c.json(latestResetJob);
-    return c.json({ error: { code: "NOT_FOUND", message: "Reset job not found" } }, 404);
-  });
-
-  routes.get("/reenrichments/jobs", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const conflict = await getRecreateConflict(db);
-    const currentReenrichJob = getCurrentReenrichJob();
-    return c.json({
-      active: currentReenrichJob !== null || isRecreateActive(),
-      currentJob: currentReenrichJob,
-      latestJob: getLatestReenrichJob(),
-      blockedBy: currentReenrichJob ? null : conflict,
-    });
-  });
-
-  routes.get("/reenrichments/jobs/:id", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const id = c.req.param("id");
-    const currentReenrichJob = getCurrentReenrichJob();
-    const latestReenrichJob = getLatestReenrichJob();
-    if (currentReenrichJob?.id === id) return c.json(currentReenrichJob);
-    if (latestReenrichJob?.id === id) return c.json(latestReenrichJob);
-    return c.json({ error: { code: "NOT_FOUND", message: "Re-enrich job not found" } }, 404);
-  });
-
-  routes.delete("/reenrichments/jobs/:id", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const id = c.req.param("id");
-    const currentReenrichJob = getCurrentReenrichJob();
-    if (!currentReenrichJob || currentReenrichJob.id !== id) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Active re-enrich job not found" } }, 404);
-    }
-    currentReenrichJob.cancelRequested = true;
-    currentReenrichJob.error = "Stop requested";
-    return c.json({ message: "Stop requested.", job: currentReenrichJob }, 202);
-  });
-
-  /**
-   * GET /api/entities/rebuilds/jobs
-   * Active/most-recent rebuild job (step 2 of two-step replay path).
-   */
-  routes.get("/rebuilds/jobs", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const conflict = await getRecreateConflict(db);
-    const currentRebuildJob = getCurrentRebuildJob();
-    return c.json({
-      active: currentRebuildJob !== null || isRecreateActive(),
-      currentJob: currentRebuildJob,
-      latestJob: getLatestRebuildJob(),
-      blockedBy: currentRebuildJob ? null : conflict,
-    });
-  });
-
-  routes.get("/rebuilds/jobs/:id", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const id = c.req.param("id");
-    const currentRebuildJob = getCurrentRebuildJob();
-    const latestRebuildJob = getLatestRebuildJob();
-    if (currentRebuildJob?.id === id) return c.json(currentRebuildJob);
-    if (latestRebuildJob?.id === id) return c.json(latestRebuildJob);
-    return c.json({ error: { code: "NOT_FOUND", message: "Rebuild job not found" } }, 404);
-  });
-
-  /**
-   * POST /api/entities/rebuilds
-   * Step 2 (replay path) of the two-step rebuild. Requires a pendingRebuildId
-   * returned by a prior /resets call. Replays existing facts only — does not
-   * call the LLM. Use /reenrichments for the re-extract path.
-   *
-   * Body: { pendingRebuildId: string }
-   */
-  routes.post("/rebuilds", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const body = (await c.req.json().catch(() => ({}))) as { pendingRebuildId?: string };
-    if (typeof body.pendingRebuildId !== "string" || body.pendingRebuildId.length === 0) {
-      return c.json({ error: { code: "BAD_REQUEST", message: "pendingRebuildId is required" } }, 400);
-    }
-    const pendingRebuildId = body.pendingRebuildId;
-
-    if (getCurrentResetJob() || getCurrentReenrichJob() || getCurrentRebuildJob()) {
-      return c.json({ error: { code: "RECREATE_ACTIVE", message: "Another job is already active" } }, 409);
-    }
-
-    const promote = promotePendingRebuild(pendingRebuildId);
-    if (promote !== "promoted") {
-      return c.json(
-        {
-          error: {
-            code: promote === "expired" ? "PENDING_REBUILD_EXPIRED" : "PENDING_REBUILD_INVALID",
-            message: `pendingRebuildId ${promote.replace(/_/g, " ")}`,
-          },
-        },
-        409,
-      );
-    }
-
-    const job = newRebuildJob({ pendingRebuildId });
-    job.phase = "replaying_facts";
-    setCurrentRebuildJob(job);
-    const triggeredByUserId = (c.get("sub") as string | undefined) ?? "system";
-
-    void (async () => {
-      try {
-        const result = await recreateEntityGraph({
-          db,
-          logger: logger.child({ jobId: job.id, component: "entity-rebuild" }),
-          triggeredByUserId,
-          skipReset: true,
-          lockAlreadyHeld: true,
-          llmPromotionThreshold: config.LLM_PROMOTION_THRESHOLD,
-          coMentionContributesToThreshold: config.CO_MENTION_CONTRIBUTES_TO_THRESHOLD,
-          onProgress: (progress) => {
-            job.progress = progress;
-          },
-        });
-        job.recreate = result;
-        job.phase = "done";
-        job.finishedAt = new Date().toISOString();
-      } catch (err) {
-        job.error = err instanceof Error ? err.message : String(err);
-        job.phase = "failed";
-        job.finishedAt = new Date().toISOString();
-        logger.error({ err, jobId: job.id }, "Rebuild job failed");
-      } finally {
-        setLatestRebuildJob(job);
-        setCurrentRebuildJob(null);
-        if (isRecreateActive()) endRecreateLock();
-      }
-    })();
-
-    return c.json(
-      {
-        message: "Rebuild started.",
-        job: { id: job.id, phase: job.phase, startedAt: job.startedAt },
-      },
-      202,
-    );
-  });
-
-  /**
-   * DELETE /api/entities/rebuilds/pending/:id
-   * Release a pending recreate lock (caller hit Cancel on step 2). Idempotent
-   * from the UI perspective:
-   *   - 204 when the lock was pending under this id and released.
-   *   - 409 when the lock was already promoted into an active rebuild.
-   *   - 404 otherwise (no pending lock or id mismatch — expired counts too).
-   */
-  routes.delete("/rebuilds/pending/:id", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const id = c.req.param("id");
-    const result = cancelPendingRebuild(id);
-    if (result === "released") return c.body(null, 204);
-    if (result === "already_promoted") {
-      return c.json({ error: { code: "PENDING_REBUILD_INVALID", message: "pendingRebuildId already promoted" } }, 409);
-    }
-    return c.json({ error: { code: "NOT_FOUND", message: `pendingRebuildId ${result.replace(/_/g, " ")}` } }, 404);
-  });
-
-  /**
-   * GET /api/entities/rebuilds/pending
-   * Inspect the current pending lock (used by the dialog when the page is
-   * refreshed mid-flow to recover step state).
-   */
-  routes.get("/rebuilds/pending", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const pending = getPendingRebuild();
-    if (!pending) return c.json({ pending: null });
-    return c.json({
-      pending: {
-        pendingRebuildId: pending.pendingRebuildId,
-        expiresAt: new Date(pending.expiresAt).toISOString(),
-        createdAt: new Date(pending.createdAt).toISOString(),
-      },
-    });
-  });
-
-  routes.post("/reenrichments", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const body = (await c.req.json().catch(() => ({}))) as {
-      scope?: ReenrichScope;
-      confirm?: string;
-      dryRun?: boolean;
-      runAfter?: boolean;
-      pendingRebuildId?: string;
-      // Guard rail: /reenrichments never accepts categories. A misrouted
-      // reset payload landing here would silently re-extract on the entire
-      // graph (the original silent-discard regression).
-      categories?: unknown;
-    };
-    if (body.categories !== undefined) {
-      return c.json(
-        { error: { code: "BAD_REQUEST", message: "/reenrichments does not accept categories; use /resets" } },
-        400,
-      );
-    }
-
-    const scope = parseReenrichScope(body.scope);
-    if (!scope) {
-      return c.json({ error: { code: "BAD_REQUEST", message: "scope is required" } }, 400);
-    }
-    const dryRun = body.dryRun === true;
-    const allScope = "all" in scope && scope.all === true;
-    if (allScope && !dryRun && body.confirm !== REENRICH_CONFIRM_TOKEN) {
-      return c.json({ error: { code: "BAD_REQUEST", message: `confirm must be ${REENRICH_CONFIRM_TOKEN}` } }, 400);
-    }
-
-    const resolved = await resolveReenrichFileIds(db, scope);
-    if (dryRun) {
-      const summary = await computeReenrichDryRun(db, resolved.fileIds, resolved.missingFileIds);
-      return c.json({ dryRun: true, ...summary });
-    }
-    if (resolved.fileIds.length === 0) {
-      return c.json(
-        { error: { code: "BAD_REQUEST", message: "No non-archived files matched the requested scope" } },
-        400,
-      );
-    }
-
-    // Two-step rebuild: the caller already holds a pending recreate lock from
-    // a prior /resets call. Promote it; only then are the standard sync/
-    // enrichment/in-flight-job conflicts checked (because our own pending
-    // lock would otherwise read as RECREATE_ACTIVE).
-    const pendingRebuildId = typeof body.pendingRebuildId === "string" ? body.pendingRebuildId : undefined;
-    let lockAlreadyHeld = false;
-    if (pendingRebuildId) {
-      const promote = promotePendingRebuild(pendingRebuildId);
-      if (promote !== "promoted") {
-        return c.json(
-          {
-            error: {
-              code: promote === "expired" ? "PENDING_REBUILD_EXPIRED" : "PENDING_REBUILD_INVALID",
-              message: `pendingRebuildId ${promote.replace(/_/g, " ")}`,
-            },
-          },
-          409,
-        );
-      }
-      lockAlreadyHeld = true;
-    } else {
-      const conflict = await getRecreateConflict(db);
-      if (conflict) {
-        return c.json({ error: { code: conflict.code, message: conflict.message } }, 409);
-      }
-    }
-    if (getCurrentResetJob()) {
-      if (lockAlreadyHeld) endRecreateLock();
-      return c.json({ error: { code: "RECREATE_ACTIVE", message: "Reset job already active" } }, 409);
-    }
-    if (getCurrentReenrichJob()) {
-      if (lockAlreadyHeld) endRecreateLock();
-      return c.json({ error: { code: "RECREATE_ACTIVE", message: "Re-enrich job already active" } }, 409);
-    }
-    if (getCurrentRebuildJob()) {
-      if (lockAlreadyHeld) endRecreateLock();
-      return c.json({ error: { code: "RECREATE_ACTIVE", message: "Rebuild job already active" } }, 409);
-    }
-
-    const runAfter = body.runAfter !== false;
-    const job = newReenrichJob({ scope, runAfter });
-    job.phase = "wiping";
-    setCurrentReenrichJob(job);
-
-    void (async () => {
-      try {
-        job.phase = "wiping";
-        const summary = await runReenrichJob({
-          db,
-          logger: logger.child({ jobId: job.id, component: "entity-reenrich" }),
-          triggeredByUserId: (c.get("sub") as string | undefined) ?? "system",
-          fileIds: resolved.fileIds,
-          missingFileIds: resolved.missingFileIds,
-          runAfter,
-          lockAlreadyHeld,
-          llmPromotionThreshold: config.LLM_PROMOTION_THRESHOLD,
-          coMentionContributesToThreshold: config.CO_MENTION_CONTRIBUTES_TO_THRESHOLD,
-          onPhase: (phase) => {
-            job.phase = phase;
-          },
-          onProgress: (progress) => {
-            job.progress = progress;
-          },
-          shouldCancel: () => job.cancelRequested === true,
-        });
-        job.summary = summary;
-        job.phase = "done";
-        job.finishedAt = new Date().toISOString();
-      } catch (err) {
-        job.error = err instanceof Error ? err.message : String(err);
-        job.phase = job.cancelRequested ? "cancelled" : "failed";
-        job.finishedAt = new Date().toISOString();
-        if (job.cancelRequested) {
-          logger.warn({ jobId: job.id }, "Re-enrich job stopped");
-        } else {
-          logger.error({ err, jobId: job.id }, "Re-enrich job failed");
-        }
-      } finally {
-        setLatestReenrichJob(job);
-        setCurrentReenrichJob(null);
-        // Promoted pending → active; runReenrichJob doesn't release the
-        // caller-held lock, so we release it here.
-        if (lockAlreadyHeld && isRecreateActive()) endRecreateLock();
-      }
-    })();
-
-    return c.json(
-      {
-        message: "Re-enrich started.",
-        files: resolved.fileIds.length,
-        missingFileIds: resolved.missingFileIds,
-        factTypes: AI_EXTRACTION_FACT_TYPES,
-        job: { id: job.id, phase: job.phase, startedAt: job.startedAt },
-      },
-      202,
-    );
-  });
-
-  /**
-   * POST /api/entities/resets
-   * Delete entities by category, optionally clearing related fact flags and
-   * rebuilding the entity graph via materialize+deterministic linking.
-   *
-   * Body: {
-   *   categories: ("manual" | "connectors" | "ai")[],
-   *   runAfter?: boolean,
-   *   dryRun?: boolean,
-   *   confirm?: string,  // required when runAfter=true OR all categories
-   *   wipeLlmFacts?: boolean,  // tombstone llm_extracted/llm_relation facts after the category purge (step-1 two-step rebuild)
-   * }
-   *
-   * When `runAfter=false`, the server begins a pending recreate lock and
-   * returns `pendingRebuildId` + `pendingRebuildExpiresAt` so the caller can
-   * present them to /rebuilds or /reenrichments for step 2.
-   */
-  routes.post("/resets", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const body = (await c.req.json().catch(() => ({}))) as {
-      categories?: string[];
-      runAfter?: boolean;
-      dryRun?: boolean;
-      confirm?: string;
-      wipeLlmFacts?: boolean;
-      // Guard rail: /resets never accepts scope. Reject explicitly so a
-      // misrouted reenrich payload doesn't silently succeed as a category
-      // reset (regression for the silent-discard bug that landed in prod).
-      scope?: unknown;
-    };
-    if (body.scope !== undefined) {
-      return c.json(
-        { error: { code: "BAD_REQUEST", message: "/resets does not accept scope; use /reenrichments" } },
-        400,
-      );
-    }
-    const categoriesIn = Array.isArray(body.categories) ? body.categories : [];
-    const categories = new Set<ResetCategory>(
-      categoriesIn.filter((c): c is ResetCategory => c === "manual" || c === "connectors" || c === "ai"),
-    );
-
-    if (categories.size === 0) {
-      return c.json({ error: { code: "BAD_REQUEST", message: "At least one category is required" } }, 400);
-    }
-
-    const runAfter = body.runAfter === true;
-    const dryRun = body.dryRun === true;
-    const wipeLlmFacts = body.wipeLlmFacts === true;
-    const allCategoriesSelected = categories.size === 3;
-
-    if ((runAfter || allCategoriesSelected) && !dryRun && body.confirm !== RESET_CONFIRM_TOKEN) {
-      return c.json({ error: { code: "BAD_REQUEST", message: `confirm must be ${RESET_CONFIRM_TOKEN}` } }, 400);
-    }
-
-    if (runAfter && categories.size === 1 && categories.has("manual")) {
-      return c.json(
-        {
-          error: { code: "BAD_REQUEST", message: "runAfter requires connectors or ai — manual has no facts to replay" },
-        },
-        400,
-      );
-    }
-
-    const includeConnectors = categories.has("connectors");
-    const includeAi = categories.has("ai");
-    const includeManual = categories.has("manual");
-
-    const factTypes = Array.from(new Set([...categories].flatMap((cat) => FACT_TYPES_BY_CATEGORY[cat] ?? [])));
-
-    if (dryRun) {
-      const counts = await computeResetCounts(db, { includeConnectors, includeAi, includeManual }, ORG_SOURCE_TYPES);
-      let dryRunFactsMarked = 0;
-      if (factTypes.length > 0) {
-        const row = await db
-          .selectFrom("indexed_file_facts")
-          .select(db.fn.countAll<number>().as("c"))
-          .where("fact_type", "in", factTypes)
-          .where("deleted_at", "is", null)
-          .where("materialized_at", "is not", null)
-          .executeTakeFirst();
-        dryRunFactsMarked = Number(row?.c ?? 0);
-      }
-      return c.json({
-        message: `Would delete ${counts.entitiesDeleted} entities.`,
-        dryRun: true,
-        entitiesDeleted: counts.entitiesDeleted,
-        candidatesCleared: counts.candidatesCleared,
-        reviewQueueCleared: counts.reviewQueueCleared,
-        reviewEvidenceCleared: counts.reviewEvidenceCleared,
-        rejectionsCleared: counts.rejectionsCleared,
-        factsMarkedUnmaterialized: dryRunFactsMarked,
-      });
-    }
-
-    const conflict = await getRecreateConflict(db);
-    if (conflict) {
-      return c.json({ error: { code: conflict.code, message: conflict.message } }, 409);
-    }
-    if (getCurrentResetJob()) {
-      return c.json({ error: { code: "RECREATE_ACTIVE", message: "Reset job already active" } }, 409);
-    }
-    if (getCurrentReenrichJob()) {
-      return c.json({ error: { code: "RECREATE_ACTIVE", message: "Re-enrich job already active" } }, 409);
-    }
-    if (getCurrentRebuildJob()) {
-      return c.json({ error: { code: "RECREATE_ACTIVE", message: "Rebuild job already active" } }, 409);
-    }
-
-    const job = newResetJob({ categories: [...categories], runAfter, wipeLlmFacts });
-    job.phase = "resetting";
-
-    // Lock acquisition:
-    //   runAfter=true  → active (legacy single-shot reset+rebuild)
-    //   runAfter=false → pending (step 1 of two-step rebuild: hold the lane
-    //                   until step 2 promotes or TTL expires)
-    let pendingRebuildId: string | null = null;
-    let pendingExpiresAtIso: string | null = null;
-    if (runAfter) {
-      beginRecreateLock();
-    } else {
-      pendingRebuildId = randomUUID();
-      const { expiresAt } = beginPendingRebuild({ pendingRebuildId });
-      pendingExpiresAtIso = new Date(expiresAt).toISOString();
-      job.pendingRebuildId = pendingRebuildId;
-      job.pendingRebuildExpiresAt = pendingExpiresAtIso;
-    }
-    setCurrentResetJob(job);
-    const triggeredByUserId = (c.get("sub") as string | undefined) ?? "system";
-
-    void (async () => {
-      try {
-        job.progress = { phase: "resetting", completed: 0, total: 1 };
-        const summary = await performReset(db, {
-          includeConnectors,
-          includeAi,
-          includeManual,
-          orgSourceTypes: ORG_SOURCE_TYPES,
-          factTypes,
-        });
-        job.reset = summary.resetSummary;
-        if (wipeLlmFacts) {
-          const tomb = await tombstoneActiveLlmFacts(db, logger.child({ jobId: job.id, phase: "tombstone-llm" }));
-          job.llmFactsWiped = tomb;
-        }
-        job.progress = { phase: "resetting", completed: 1, total: 1 };
-        if (!runAfter) {
-          job.phase = "done";
-          job.finishedAt = new Date().toISOString();
-          return;
-        }
-        job.phase = "replaying_facts";
-        const result = await recreateEntityGraph({
-          db,
-          logger: logger.child({ jobId: job.id }),
-          triggeredByUserId,
-          skipReset: true,
-          lockAlreadyHeld: true,
-          llmPromotionThreshold: config.LLM_PROMOTION_THRESHOLD,
-          coMentionContributesToThreshold: config.CO_MENTION_CONTRIBUTES_TO_THRESHOLD,
-          materializeFactTypes: factTypes.length > 0 ? factTypes : undefined,
-          onProgress: (progress) => {
-            job.progress = progress;
-          },
-        });
-        job.replay = result.replay;
-        job.recreate = result;
-        job.phase = "done";
-        job.finishedAt = new Date().toISOString();
-      } catch (err) {
-        job.error = err instanceof Error ? err.message : String(err);
-        job.phase = "failed";
-        job.finishedAt = new Date().toISOString();
-        logger.error({ err, jobId: job.id, runAfter, wipeLlmFacts }, "Reset job failed");
-        // When the step-1 reset itself fails, release the pending lock so sync
-        // and enrichment aren't stuck behind a graph the operator never got.
-        if (!runAfter && pendingRebuildId) {
-          cancelPendingRebuild(pendingRebuildId);
-        }
-      } finally {
-        setLatestResetJob(job);
-        setCurrentResetJob(null);
-        // runAfter=true: release the active lock we held.
-        // runAfter=false: pending lock stays for step 2 (or TTL releases it).
-        if (runAfter && isRecreateActive()) endRecreateLock();
-      }
-    })();
-
-    return c.json(
-      {
-        message: runAfter ? "Reset started, rebuild will follow." : "Reset started.",
-        job: {
-          id: job.id,
-          phase: job.phase,
-          startedAt: job.startedAt,
-          ...(pendingRebuildId ? { pendingRebuildId, pendingRebuildExpiresAt: pendingExpiresAtIso ?? undefined } : {}),
-        },
-        ...(pendingRebuildId ? { pendingRebuildId, pendingRebuildExpiresAt: pendingExpiresAtIso } : {}),
-      },
-      202,
-    );
-  });
-
-  /**
-   * DELETE /api/entities/tentative
-   * Delete all tentative entities and their mentions.
-   * Must be registered before /:id to prevent "tentative" matching as an ID.
-   */
-  routes.delete("/tentative", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const typeFilter = c.req.query("type")?.split(",").filter(Boolean);
-
-    let query = db.selectFrom("entities").select("id").where("status", "=", "tentative");
-    if (typeFilter && typeFilter.length > 0) {
-      query = query.where("source_type", "in", typeFilter);
-    }
-    const entities = await query.execute();
-
-    if (entities.length === 0) {
-      return c.json({ message: "No tentative entities to delete.", count: 0 });
-    }
-
-    const ids = entities.map((e) => e.id);
-    await db.deleteFrom("entities").where("id", "in", ids).execute();
-
-    return c.json({
-      message: `Deleted ${ids.length} tentative entities and their mentions.`,
-      count: ids.length,
-    });
-  });
-
-  /**
-   * GET /api/entities/:id
-   *
-   * Backwards-compatible: existing callers keep getting `{ entity, sourceRefs }`
-   * with the same field names. The drawer surface receives an additional
-   * `entity.profile` block (aggregates + a deterministic summary) so the
-   * open-feel is a single round-trip. The profile block is always present so
-   * the web layer can rely on it; the experimental flag gates the drawer UI,
-   * not the response shape.
-   */
-  routes.get("/:id", async (c) => {
-    const entity = await repo.getEntity(c.req.param("id"));
-    if (!entity) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found" } }, 404);
-    }
-
-    const [sourceRefs, facts] = await Promise.all([
-      db.selectFrom("entity_source_refs").selectAll().where("entity_id", "=", entity.id).execute(),
-      loadEntityFactsForId(entity.id),
-    ]);
-    if (!facts) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found" } }, 404);
-    }
-
-    const parsedAliases = entity.aliases ? (JSON.parse(entity.aliases) as string[]) : [];
-    const viewer = getFileViewer(c);
-    const activity = await loadActivityStats(entity.id, viewer);
-    const summary = buildSummary(facts, activity);
-
-    return c.json({
-      entity: {
-        id: entity.id,
-        name: entity.name,
-        sourceType: entity.source_type,
-        subtype: entity.subtype,
-        aliases: parsedAliases,
-        metadata: facts.metadata,
-        status: entity.status,
-        hotness: entity.hotness,
-        createdAt: entity.created_at,
-        updatedAt: entity.updated_at,
-        profile: {
-          entityType: facts.entityType,
-          mentionCount: facts.mentionCount,
-          sourceCounts: facts.sourceCounts,
-          firstSeenAt: facts.firstSeenAt,
-          lastSeenAt: facts.lastSeenAt,
-          domainsForCompany: facts.domainsForCompany,
-          summary,
-        },
-      },
-      sourceRefs: sourceRefs.map((r) => ({
-        id: r.id,
-        source: r.source,
-        sourceId: r.source_id,
-        sourceUrl: r.source_url,
-        lastSeenAt: r.last_seen_at,
-      })),
-    });
-  });
-
-  /**
-   * GET /api/entities/:id/relations
-   * Drawer Relationships section. Caps each side at 200; sets `truncated`
-   * when more rows exist. Gated by EXPERIMENTAL_FLAG (drawer UI is flagged).
-   */
-  routes.get("/:id/relations", async (c) => {
-    if (!config.EXPERIMENTAL_FLAG) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Not found" } }, 404);
-    }
-    const entity = await repo.getEntity(c.req.param("id"));
-    if (!entity) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found" } }, 404);
-    }
-    const result = await relRepo.listRelationsForEntity(entity.id, { limit: RELATIONS_LIMIT });
-    return c.json({
-      outgoing: result.outgoing.sort(relationCompare),
-      incoming: result.incoming.sort(relationCompare),
-      truncated: result.truncated,
-      totalCount: result.totalCount,
-    });
-  });
-
-  /**
-   * GET /api/entities/:id/relations/:rid/evidence
-   * Evidence rows for a single relation, filtered by file-level RBAC.
-   * `visibleCount < totalCount` is the "N more not visible to you" signal —
-   * we expose that hidden evidence exists, not its content (file names and
-   * snippets are filtered out for hidden rows).
-   */
-  routes.get("/:id/relations/:rid/evidence", async (c) => {
-    if (!config.EXPERIMENTAL_FLAG) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Not found" } }, 404);
-    }
-    const entityId = c.req.param("id");
-    const relationshipId = c.req.param("rid");
-    const entity = await repo.getEntity(entityId);
-    if (!entity) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found" } }, 404);
-    }
-    const relation = await relRepo.getRelationship(relationshipId);
-    if (!relation || (relation.source_entity_id !== entityId && relation.target_entity_id !== entityId)) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Relation not found" } }, 404);
-    }
-    const viewer = getFileViewer(c);
-    const result = await relRepo.listEvidenceForRelation(relationshipId, { limit: EVIDENCE_LIMIT, viewer });
-    return c.json(result);
-  });
-
-  /**
-   * GET /api/entities/:id/timeline
-   * File mentions for an entity, grouped by month newest-first. File RBAC
-   * applied; capped at 100 visible rows.
-   */
-  routes.get("/:id/timeline", async (c) => {
-    if (!config.EXPERIMENTAL_FLAG) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Not found" } }, 404);
-    }
-    const entity = await repo.getEntity(c.req.param("id"));
-    if (!entity) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found" } }, 404);
-    }
-    const viewer = getFileViewer(c);
-    const result = await timelineRepo.listTimelineForEntity(entity.id, { limit: TIMELINE_LIMIT, viewer });
-    return c.json(result);
-  });
-
-  /**
-   * PATCH /api/entities/:id
-   * Update entity name, source_type, status, or aliases.
-   */
-  routes.patch("/:id", async (c) => {
-    const entity = await repo.getEntity(c.req.param("id"));
-    if (!entity) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found" } }, 404);
-    }
-
-    const body = (await c.req.json()) as {
-      name?: string;
-      sourceType?: string;
-      status?: string;
-      aliases?: string[];
-    };
-
-    const updates: Record<string, unknown> = {};
-    if (body.name !== undefined) updates.name = body.name;
-    if (body.sourceType !== undefined) updates.source_type = body.sourceType;
-    if (body.status !== undefined) updates.status = body.status;
-    if (body.aliases !== undefined) updates.aliases = JSON.stringify(body.aliases);
-
-    if (Object.keys(updates).length > 0) {
-      await repo.updateEntity(entity.id, updates);
-    }
-
-    const updated = await repo.getEntity(entity.id);
-    if (!updated) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found after update" } }, 404);
-    }
-    return c.json({
-      entity: {
-        id: updated.id,
-        name: updated.name,
-        sourceType: updated.source_type,
-        subtype: updated.subtype,
-        aliases: updated.aliases ? JSON.parse(updated.aliases) : [],
-        status: updated.status,
-      },
-    });
-  });
-
-  /**
-   * DELETE /api/entities/:id
-   * Delete an entity and its mentions/source refs (cascade).
-   */
-  routes.delete("/:id", async (c) => {
-    const denied = denyIfNotAdmin(c);
-    if (denied) return denied;
-    const entity = await repo.getEntity(c.req.param("id"));
-    if (!entity) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found" } }, 404);
-    }
-
-    await db.deleteFrom("entities").where("id", "=", entity.id).execute();
-    return c.json({ success: true });
-  });
-
-  /**
-   * GET /api/entities/:id/mentions
-   *   ?source=clickup,fireflies       — filter by indexed_file source
-   *   &since=2026-03-01               — date filter
-   *   &limit=20&offset=0
-   */
-  routes.get("/:id/mentions", async (c) => {
-    const entityId = c.req.param("id");
-    const sourceFilter = c.req.query("source")?.split(",").filter(Boolean);
-    const since = c.req.query("since");
-    const limit = Math.min(Number(c.req.query("limit")) || 20, 100);
-    const offset = Number(c.req.query("offset")) || 0;
-    const viewer = getFileViewer(c);
-
-    const entity = await repo.getEntity(entityId);
-    if (!entity) {
-      return c.json({ error: { code: "NOT_FOUND", message: "Entity not found" } }, 404);
-    }
-
-    let query = db
-      .selectFrom("entity_mentions")
-      .innerJoin("indexed_files", "indexed_files.id", "entity_mentions.indexed_file_id")
-      .select([
-        "entity_mentions.id",
-        "entity_mentions.context_snippet",
-        "entity_mentions.chunk_index",
-        "entity_mentions.mentioned_at",
-        "indexed_files.id as file_id",
-        "indexed_files.file_name",
-        "indexed_files.file_type",
-        "indexed_files.source",
-        "indexed_files.source_path",
-        "indexed_files.provider_url",
-        "indexed_files.source_updated_at",
-        "indexed_files.source_created_at",
-      ])
-      .where("entity_mentions.entity_id", "=", entityId)
-      .orderBy(
-        sql`COALESCE(indexed_files.source_updated_at, indexed_files.source_created_at, entity_mentions.mentioned_at)`,
-        "desc",
-      );
-
-    if (sourceFilter && sourceFilter.length > 0) {
-      query = query.where("indexed_files.source", "in", sourceFilter);
-    }
-    if (since) {
-      query = query.where(
-        sql`COALESCE(indexed_files.source_updated_at, indexed_files.source_created_at, entity_mentions.mentioned_at)`,
-        ">=",
-        since,
-      );
-    }
-    if (!viewer.isAdmin) {
-      query = query.where(fileVisibilityPredicate(viewer));
-    }
-
-    query = query.limit(limit).offset(offset);
-    const mentions = await query.execute();
-
-    const buildCountQuery = (gated: boolean) => {
-      let q = db
-        .selectFrom("entity_mentions")
-        .innerJoin("indexed_files", "indexed_files.id", "entity_mentions.indexed_file_id")
-        .select(sql<number>`count(entity_mentions.id)`.as("total"))
-        .where("entity_mentions.entity_id", "=", entityId);
-      if (sourceFilter && sourceFilter.length > 0) {
-        q = q.where("indexed_files.source", "in", sourceFilter);
-      }
-      if (since) {
-        q = q.where(
-          sql`COALESCE(indexed_files.source_updated_at, indexed_files.source_created_at, entity_mentions.mentioned_at)`,
-          ">=",
-          since,
-        );
-      }
-      if (gated) q = q.where(fileVisibilityPredicate(viewer));
-      return q;
-    };
-
-    const visibleCount = Number((await buildCountQuery(!viewer.isAdmin).executeTakeFirst())?.total ?? 0);
-    const hiddenCount = viewer.isAdmin
-      ? 0
-      : Math.max(0, Number((await buildCountQuery(false).executeTakeFirst())?.total ?? 0) - visibleCount);
-
-    return c.json({
-      mentions: mentions.map((m) => ({
-        id: m.id,
-        contextSnippet: m.context_snippet,
-        chunkIndex: m.chunk_index,
-        mentionedAt: m.mentioned_at,
-        sourceDate: m.source_updated_at ?? m.source_created_at ?? m.mentioned_at,
-        file: {
-          id: m.file_id,
-          fileName: m.file_name,
-          fileType: m.file_type,
-          source: m.source,
-          sourcePath: m.source_path,
-          providerUrl: m.provider_url,
-        },
-      })),
-      total: visibleCount,
-      hiddenCount,
-    });
-  });
-
-  /**
-   * POST /api/entities/backfill
-   * Mark enriched files that have no entity mentions for re-enrichment.
-   * The next enrichment run will process them with entity linking enabled.
-   * Optional: ?source=clickup,fireflies to limit to specific sources.
-   */
-  routes.post("/backfill", async (c) => {
-    const sourceFilter = c.req.query("source")?.split(",").filter(Boolean);
-
-    // Find files that are enriched but have no entity mentions
-    let query = db
-      .selectFrom("indexed_files")
-      .select(["indexed_files.id"])
-      .where("indexed_files.embedding_status", "=", "done")
-      .where("indexed_files.is_archived", "=", 0)
-      .where("indexed_files.id", "not in", db.selectFrom("entity_mentions").select("indexed_file_id"));
-
-    if (sourceFilter && sourceFilter.length > 0) {
-      query = query.where("indexed_files.source", "in", sourceFilter);
-    }
-
-    const files = await query.execute();
-
-    if (files.length === 0) {
-      return c.json({ message: "No files need entity backfill.", count: 0 });
-    }
-
-    // Reset their embedding_status to pending so enrichment picks them up
-    const fileIds = files.map((f) => f.id);
-    await db.updateTable("indexed_files").set({ embedding_status: "pending" }).where("id", "in", fileIds).execute();
-
-    return c.json({
-      message: `Marked ${fileIds.length} files for re-enrichment with entity linking.`,
-      count: fileIds.length,
-    });
-  });
-
+  routes.route("/", createEntityMaintenanceRoutes(db, deps));
+  routes.route("/", createEntityProfileRoutes(db, deps));
   return routes;
 }
